% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentBalancingMethod.R
\name{cbm.sort}
\alias{cbm.sort}
\title{Sort the given items by their value to the simulee, with the best item first.}
\usage{
cbm.sort(eligibleModuleIndices, simulee, simulation)
}
\arguments{
\item{eligibleModuleIndices}{A vector of item module indices, which are eligible for selection.}

\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
The eligibleModuleIndices, sorted by their value to the simulee.
}
\description{
Sort the given items by their value to the simulee, with the best item first.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByNormal(1)(), simulation)
  eligibleModuleIndices = 1:nrow(simulation$modules)
  balancedModuleIndices = cbm.sort(eligibleModuleIndices, simulee, simulation)
}
