% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentBalancingMethod.R
\name{cbm.wpm1.sort}
\alias{cbm.wpm1.sort}
\title{Sort the given items by their value to the simulee according to the Weighted Penalty Method (wpm) algorithm, with the best item first.}
\usage{
cbm.wpm1.sort(eligibleModuleIndices, simulee, simulation)
}
\arguments{
\item{eligibleModuleIndices}{A vector of itempool indices eligible for selection.}

\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
The eligibleModuleIndices, sorted by their value to the simulee.
}
\description{
The `wpm1` algorithm is the same algorithm as `wpm`, but optimized to run faster for tests containing only discrete single item modules.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByNormal(1)(), simulation)
  eligibleModuleIndices = 1:nrow(simulation$modules)
  balancedModuleIndices = cbm.wpm.sort(eligibleModuleIndices, simulee, simulation)
}
\seealso{
cbm.wpm.sort for general details.
}
