% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abilityEstimator.R
\name{estimateAbility.eap}
\alias{estimateAbility.eap}
\title{Calculate an IRT EAP ability estimate for the given sequence of scores and items.}
\usage{
estimateAbility.eap(item_pars, scores, simulation)
}
\arguments{
\item{item_pars}{a matrix of doubles, items by params. Must have columns for $PAR_* matching the $MODEL}

\item{scores}{a vector of integers. A score for each item.  Must be same length as nrow(item_pars).}

\item{simulation}{an object defining the test that is being run.}
}
\value{
a list with properties theta and csem.
}
\description{
Calculate an IRT EAP ability estimate for the given sequence of scores and items.
}
\examples{
  simulation = readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator"))
  item_pars = as.matrix(simulation$itempool[1:5,c("PAR_1","PAR_2","PAR_3")])
  scores = generateScores(item_pars, -0.4)
  abilityEstimate = estimateAbility.eap(item_pars, scores, simulation)
  abilityEstimate.irtoys = irtoys::eap(resp=scores, ip=list(est=item_pars), qu=irtoys::normal.qu(51, -5.0, 5.0, 0.0, 1.0, scaling=""))
}
