% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{generateSimuleesByNormal}
\alias{generateSimuleesByNormal}
\title{Return a generator function that will generate one simulee each time it is called.
The true theta of the generated simulees will follow a normal distribution.
Once the specified number of simulees has been reached, no more simulees will be generated.}
\usage{
generateSimuleesByNormal(
  numSimulees,
  mean = 0,
  sd = 1,
  randomSeeds = NULL,
  min_value = -Inf,
  max_value = Inf
)
}
\arguments{
\item{numSimulees}{The number of simulees that should be generated.}

\item{mean}{The mean of the true theta for the generated simulees.}

\item{sd}{The standard deviation of the true theta for the generated simulees.}

\item{randomSeeds}{A vector of integers. The source of random seeds for each simulee.}
}
\value{
A generator function that will generate one simulee each time it is called.
}
\description{
Return a generator function that will generate one simulee each time it is called.
The true theta of the generated simulees will follow a normal distribution.
Once the specified number of simulees has been reached, no more simulees will be generated.
}
\examples{
  simuleeGenerator = generateSimuleesByNormal(numSimulees = 5, mean = 0.5, sd = 1.2, randomSeeds = 10001:10003)
  simulee = simuleeGenerator()
}
