% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{generateSimuleesByTrueTheta}
\alias{generateSimuleesByTrueTheta}
\title{Return a generator function that will generate one simulee each time it is called.
The true theta of the generated simulees will come from the next value in the given vector.
Once all true thetas have been consumed, no more simulees will be generated.}
\usage{
generateSimuleesByTrueTheta(trueThetas, randomSeeds = NULL)
}
\arguments{
\item{trueThetas}{A vector of doubles. The source of true thetas for each simulee.}

\item{randomSeeds}{A vector of integers. The source of random seeds for each simulee.}
}
\value{
A generator function that will generate one simulee each time it is called.
}
\description{
Return a generator function that will generate one simulee each time it is called.
The true theta of the generated simulees will come from the next value in the given vector.
Once all true thetas have been consumed, no more simulees will be generated.
}
\examples{
  simuleeGenerator = generateSimuleesByTrueTheta(-2:2, 10001:10003)
  simulee = simuleeGenerator()
}
