% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{getResult.IEC}
\alias{getResult.IEC}
\title{Calculate stats analyzing the quality of the Item Exposure Control aspect of
a completed simulation.}
\usage{
getResult.IEC(simulation, output)
}
\arguments{
\item{simulation}{An object defining the test that was run.}

\item{output}{A tibble of data from simulees that have completed the simulation.}
}
\value{
An object containing multiple stats and tables showing the quality
  of the completed simulation.
}
\description{
Calculate stats analyzing the quality of the Item Exposure Control aspect of
a completed simulation.
}
\examples{
  simulation = readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator"))
  result = runSimulation(simulation, generateSimuleesByNormal(5))
  result$iec = getResult.IEC(simulation, result$output)
}
