% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{getResult.IRT}
\alias{getResult.IRT}
\title{Calculate stats analyzing the quality of the IRT Ability Estimation aspect
of a completed simulation.}
\usage{
getResult.IRT(simulation, output)
}
\arguments{
\item{simulation}{An object defining the test that was run.}

\item{output}{A tibble of data from simulees that have completed the simulation.}
}
\value{
An object containing multiple stats and tables showing the quality
  of the completed simulation.
}
\description{
Calculate stats analyzing the quality of the IRT Ability Estimation aspect
of a completed simulation.
}
\examples{
  simulation = readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator"))
  result = runSimulation(simulation, generateSimuleesByNormal(5))
  result$irt = getResult.IRT(simulation, result$output)
}
