% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemExposureControl.R
\name{iec.filter}
\alias{iec.filter}
\title{Filter the given items down to the best N rows, as defined by the simulation.}
\usage{
iec.filter(balancedModuleIndices, simulation)
}
\arguments{
\item{balancedModuleIndices}{A vector of indices, which are candidates for selection sorted by their value to the simulee.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
The balancedModuleIndices, filtered to the best N values.
}
\description{
Filter the given items down to the best N rows, as defined by the simulation.
}
\examples{
  simulation = readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator"))
  balancedModuleIndices = sample(1:nrow(simulation$modules), 4)
  candidateModuleIndices = iec.filter(balancedModuleIndices, simulation)
}
