% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemExposureControl.R
\name{iec.select}
\alias{iec.select}
\title{Select a single index randomly from a given vector of indices.}
\usage{
iec.select(candidateModuleIndices)
}
\arguments{
\item{candidateModuleIndices}{A vector of indices, which are candidates for selection.}
}
\value{
The single randomly chosen index from candidateModuleIndices.
}
\description{
Select a single index randomly from a given vector of indices.
}
\examples{
  candidateModuleIndices = sample(1:9, 4)
  selectedIndex = iec.select(candidateModuleIndices)
}
