% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{initSimulation}
\alias{initSimulation}
\title{initSimulation}
\usage{
initSimulation(simulation)
}
\description{
initSimulation
}
\examples{
  options(warn=-1)
  simulation = readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator"))
  simulation = initSimulation(simulation)
}
