% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{initSimulee}
\alias{initSimulee}
\title{initSimulee}
\usage{
initSimulee(simulee, simulation)
}
\description{
initSimulee
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  simulee = generateSimuleesByTrueTheta(-2, 10001)()
  simulee = initSimulee(simulee, simulation)
}
