% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.adaptive.select}
\alias{isr.adaptive.select}
\title{Returns the next items that should be assigned to the simulee, selected
from the given simulation.}
\usage{
isr.adaptive.select(simulee, simulation)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
A list containing the selected moduleIndex, plus other values useful for understanding the selection.
}
\description{
The `adaptive` algorithm uses content balancing and item exposure algorithms
to select the best item in the item pool for the given simulee.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  selection = isr.adaptive.select(simulee, simulation)
}
