% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.checkIncompletePassage}
\alias{isr.checkIncompletePassage}
\title{isr.checkIncompletePassage}
\usage{
isr.checkIncompletePassage(
  prevSlot,
  eligibleModuleIndices,
  simulee,
  simulation
)
}
\description{
isr.checkIncompletePassage
}
\examples{
  options(warn=-1)
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  eligibleModuleIndices = NULL
  prevSlot = 1
  prevSlotPsgId = isr.checkIncompletePassage(prevSlot, simulee, simulation)
}
