% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.filterByConstraintControl}
\alias{isr.filterByConstraintControl}
\title{isr.filterByConstraintControl}
\usage{
isr.filterByConstraintControl(
  eligibleModuleIndices,
  incompletePsgId,
  slot,
  simulation
)
}
\description{
isr.filterByConstraintControl
}
\examples{
  options(warn=-1)
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  slot = 1
  incompletePsgId = NULL
  eligibleModuleIndices = 1:nrow(simulation$modules)
  filteredIndices = isr.filterByConstraintControl(eligibleModuleIndices, incompletePsgId, slot, simulation)
}
