% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.getEligibleModules}
\alias{isr.getEligibleModules}
\title{isr.getEligibleModules}
\usage{
isr.getEligibleModules(slot, blocksReleased, simulee, simulation)
}
\description{
isr.getEligibleModules
}
\examples{
  options(warn=-1)
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  slot = 1
  blocksReleased = TRUE
  eligibleModuleIndices = isr.getEligibleModules(slot, blocksReleased, simulee, simulation)
}
