% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.linear.select}
\alias{isr.linear.select}
\title{Returns the next items that should be assigned to the simulee, selected
from the given simulation.}
\usage{
isr.linear.select(simulee, simulation)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
A list containing the selected moduleIndex, plus other values useful for understanding the selection.
}
\description{
The `linear` algorithm return the first unassigned item in the item pool.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-linear.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  selection = isr.linear.select(simulee, simulation)
}
