% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.multistage.select}
\alias{isr.multistage.select}
\title{Returns the next items that should be assigned to the simulee, selected
from the given simulation.}
\usage{
isr.multistage.select(simulee, simulation, stage)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}

\item{stage}{A number for the stage of the test to be selected}
}
\value{
A list containing the selected moduleIndex, plus other values useful for understanding the selection.
}
\description{
The `multistage` algorithm selects the item module with the best information for
the current stage, and return all the items within that module
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/internal/shallow-multistage.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  stage = 1
  selection = isr.multistage.select(simulee, simulation, stage)
}
