% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.optimal.select}
\alias{isr.optimal.select}
\title{Returns the next items that should be assigned to the simulee, selected
from the given simulation.}
\usage{
isr.optimal.select(simulee, simulation)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
A list containing the selected moduleIndex, plus other values useful for understanding the selection.
}
\description{
The `optimal` algorithm uses information value of each item as the objective,
then uses lpSolveAPI to find a complete test solution that maximizes overall test
information while still meeting blueprint constraints.  Then it selects the max
information item from the unassigned items in the solution.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  simulation$control$solver = list(name = "cbc", external = F, mipGap = 0.0001, timeout = 1000, verbose = T)
  simulation$control$solver = list(name = "gurobi", external = F, mipGap = 0.0001, timeout = 1000, verbose = T)
  simulee = generateSimuleesByTrueTheta(-2, 10001)()
  selection = isr.optimal.select(simulee, simulation)
}
