% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemSelectionRule.R
\name{isr.select}
\alias{isr.select}
\title{Returns the next items that should be assigned to the simulee, selected
from the given simulation.}
\usage{
isr.select(simulee, simulation, stage)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}

\item{stage}{A number for the stage of the test to be selected}
}
\value{
A list containing the selected moduleIndex, plus other values useful for understanding the selection.
}
\description{
Returns the next items that should be assigned to the simulee, selected
from the given simulation.
}
\examples{
  simulation = initSimulation(readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator")))
  simulee = initSimulee(generateSimuleesByTrueTheta(-2, 10001)(), simulation)
  stage = 1
  selection = isr.select(simulee, simulation, stage)
}
