% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{itemInformation.3pl}
\alias{itemInformation.3pl}
\title{Calculate the information value of each item, based on the difficulty of the item vs the theta ability level.}
\usage{
itemInformation.3pl(item_pars, theta)
}
\arguments{
\item{item_pars}{a matrix of doubles, items by params. Must have columns for $PAR_* matching the $MODEL}

\item{theta}{a double. An ability level}
}
\value{
a vector of doubles. An information value for each item based on the given theta ability
}
\description{
Calculate the information value of each item, based on the difficulty of the item vs the theta ability level.
}
\examples{
  simulation = readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator"))
  item_pars = as.matrix(simulation$itempool[,startsWith(colnames(simulation$itempool), "PAR_")])
  dimnames(item_pars) = list(simulation$itempool$ITEM_ID, colnames(simulation$itempool)[startsWith(colnames(simulation$itempool), "PAR_")])
  theta = 2.0
  itemInf = itemInformation.3pl(item_pars, theta)
  options(digits=8, scipen=8)
  > itemInf
  item042      item043      item044      item045      item048      item049      item050      item051      item052      item053
  0.0573824324 0.1207351701 0.0397438702 0.1975182873 0.1767498486 0.3782198973 0.0579802303 0.4984781223 0.2398761585 0.2363306368
  item054      item055      item056      item057      item058      item059      item060      item063      item064      item065
  0.2150341441 0.0508455836 0.0331527420 0.1038234622 0.0574888741 0.5218887394 0.0457055147 0.0293236847 0.4238179761 0.1010473900
  item066      item067      item068      item069      item070      item071      item072      item073      item075      item077
  0.0442108431 0.1380087003 0.5188451693 0.1325508768 0.2427019647 0.0326758174 0.0362685411 0.7360391894 0.1957009888 0.0401299204
  item078      item079      item080      item081      item082      item083      item084      item086      item087      item088
  0.4182005072 0.0882336121 0.0735453800 0.0284117031 0.0472563735 0.4007789174 0.5844731126 0.1430387814 0.3377729572 0.0288792556
  item089      item090      item091      item092      item093      item095      item096      item097      item098      item099
  0.0618553272 0.0234041262 0.0532715538 0.0191953506 0.0440831510 0.1729177177 0.1811829028 0.0750468411 0.0288081602 0.0430959982
  item100 ...277 total values
}
