% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{itemInformation}
\alias{itemInformation}
\title{Calculate the information value of each item, based on the difficulty of the item vs the theta ability level.}
\usage{
itemInformation(item_pars, theta)
}
\arguments{
\item{item_pars}{a matrix of doubles, items by params. Must have columns for $PAR_* matching the $MODEL}

\item{theta}{a double. An ability level}
}
\value{
a vector of doubles. An information value for each item based on the given theta ability
}
\description{
Calculate the information value of each item, based on the difficulty of the item vs the theta ability level.
}
\examples{
  simulation = readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator"))
  item_pars = as.matrix(simulation$itempool[,startsWith(colnames(simulation$itempool), "PAR_")])
  dimnames(item_pars) = list(simulation$itempool$ITEM_ID, colnames(simulation$itempool)[startsWith(colnames(simulation$itempool), "PAR_")])
  theta = 2.0
  itemInf = itemInformation(item_pars, theta)
  options(digits=8, scipen=8)
  > itemInf
  item001      item002      item003      item004      item005      item006      item007      item008      item009      item010
  0.1068118604 0.0998542776 0.4043015953 0.3817266261 0.3456384832 0.3908781653 0.8798895310 1.3384592783 0.3008433423 0.3069516829
  item011      item012      item013      item014      item015      item016      item017      item018      item019      item020
  0.0986341211 0.0484800772 0.1874231844 0.0810430934 0.2301688162 0.0886660486 0.0707584918 0.1133582940 0.5546982004 0.0746211906
  item021      item022      item023      item024      item025      item026      item027      item028      item029      item030
  0.1027127914 0.1285767322 0.1009738653 0.1600664222 0.2461996880 0.0114159873 0.9977338491 1.6097182929 0.1077239682 0.3897308353
  item031      item032      item033      item034      item035      item036      item037      item038      item039      item040
  0.5815460829 0.2585189679 0.1117031765 0.0489117847 0.0883738236 0.0372521735 0.2572516322 0.6930609211 0.9941486814 0.1084227738
  item041      item042      item043      item044      item045      item046      item047      item048      item049      item050
  0.1250331028 0.0573824324 0.1207351701 0.0397438702 0.1975182873 0.5563700696 0.1703637978 0.1767498486 0.3782198973 0.0579802303
  item051 ...391 total values
}
