% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_content_constraint}
\alias{read_content_constraint}
\title{Read in constraint.csv as content_constraint tibble}
\usage{
read_content_constraint(constraint_file)
}
\description{
Read in constraint.csv as content_constraint tibble
}
\examples{
  constraint_file = "inst/input_files/many_panels/constraint.csv"
  content_constraint = read_content_constraint(constraint_file)
}
