% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_item}
\alias{read_item}
\title{Read in item.csv as item tibble}
\usage{
read_item(item_file, item_constraint_file = NULL, content_constraint = NULL)
}
\description{
Read in item.csv as item tibble
}
\examples{
  constraint_file = "inst/input_files/many_panels/constraint.csv"
  content_constraint = read_content_constraint(constraint_file)

  item_file = "inst/input_files/many_panels/metadata.csv"
  item_constraint_file = "inst/input_files/many_panels/item_constraint.csv"
  item = read_item(item_file, item_constraint_file, content_constraint)
}
