% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_module}
\alias{read_module}
\title{Read in module.csv as module tibble}
\usage{
read_module(module_file, item = NULL)
}
\description{
Read in module.csv as module tibble
}
\examples{
  item_file = "inst/input_files/many_panels/metadata.csv"
  item = read_item(item_file)

  module_file = "inst/input_files/many_panels/module.csv"
  module = read_module(module_file, item)
}
