% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_panel}
\alias{read_panel}
\title{Read in panel.csv as panel tibble}
\usage{
read_panel(panel_file, module = NULL)
}
\description{
Read in panel.csv as panel tibble
}
\examples{
  module_file = "inst/input_files/many_panels/module.csv"
  module = read_module(module_file)

  panel_file = "inst/input_files/many_panels/panel.csv"
  panel = read_panel(panel_file, module)
}
