% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_simulation}
\alias{read_simulation}
\title{Read in test_config.json and all input csvs to instantiate a simulation object}
\usage{
read_simulation(test_config_file)
}
\description{
Read in test_config.json and all input csvs to instantiate a simulation object
}
\examples{
  test_config_file = "inst/input_files/many_panels/test.json"
  simulation = read_simulation(test_config_file)
}
