% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_test_config}
\alias{read_test_config}
\title{Read in test_config.json}
\usage{
read_test_config(test_config_file)
}
\description{
Read in test_config.json
}
\examples{
  test_config_file = "inst/input_files/many_panels/test.json"
  test_config = read_test_config(test_config_file)
}
