% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{runSimulee}
\alias{runSimulee}
\title{Run a single simulee through the given simulation.}
\usage{
runSimulee(simulee, simulation)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test to be run.}
}
\value{
An updated version of the simulee data.frame, with the test completed.
}
\description{
Run a single simulee through the given simulation.
}
\examples{
  options(warn=-1)
  simulation = initSimulation(read_simulation("inst/input_files/many_panels/test.json"))
  simulation$control$solver = list(name = "cbc", external = T, mipGap = 0.0001, timeout = 1000, verbose = T)
  simulee = generateSimuleesByTrueTheta(-2, 10001)()
  simuleeOut = runSimulee(simulee, simulation)
}
