% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemResponseTheory.R
\name{scoreProbability.theta.3pl}
\alias{scoreProbability.theta.3pl}
\title{Calculate the probability of each possible score for the given 3PL items and a specific theta}
\usage{
scoreProbability.theta.3pl(item_pars, theta)
}
\arguments{
\item{item_pars}{a matrix of doubles, items by params. Must have columns for $PAR_* matching the $MODEL}

\item{theta}{a double. An ability level}
}
\value{
a matrix of doubles, items by scores, containing the probability of each possible score for the given items and a specific theta
}
\description{
Calculate the probability of each possible score for the given 3PL items and a specific theta
}
\examples{
  simulation = readRDS(system.file("example/passage-optimal.rds", package = "CATSimulator"))
  item_pars = as.matrix(simulation$itempool[,startsWith(colnames(simulation$itempool), "PAR_")])
  dimnames(item_pars) = list(simulation$itempool$ITEM_ID, colnames(simulation$itempool)[startsWith(colnames(simulation$itempool), "PAR_")])
  theta = 2.0
  scoreProb = scoreProbability.theta.3pl(item_pars, theta)
  options(digits=8, scipen=8)
  > scoreProb       0          1
  item042 0.649862455 0.35013754
  item043 0.688542983 0.31145702
  item044 0.290662804 0.70933720
  item045 0.234881989 0.76511801
  item048 0.576947540 0.42305246
  item049 ...277 total items/rows
}
