% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addContentConstraints}
\alias{solver.addContentConstraints}
\title{Adds content constraints to the given lp, based on the given content constraints table}
\usage{
solver.addContentConstraints(lp, modules, contentCons)
}
\description{
Adds content constraints to the given lp, based on the given content constraints table
}
\examples{
  modules = simulation$modules
  contentCons = simulation$constraints$content
}
