% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addEnemyCodeConstraints}
\alias{solver.addEnemyCodeConstraints}
\title{Adds enemy code constraints to the given lp, based on the given modules}
\usage{
solver.addEnemyCodeConstraints(lp, modules, enemyCodes)
}
\description{
Adds enemy code constraints to the given lp, based on the given modules
}
\examples{
  modules = simulation$modules
  enemyCodes = sort(unique(unlist(modules$ENEMY_CODES, recursive = T)))
}
