% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addPassageConstraints}
\alias{solver.addPassageConstraints}
\title{Adds passage constraints to the given lp, based on the given passage constraints table}
\usage{
solver.addPassageConstraints(
  lp,
  modules,
  passageCons,
  minPassages = NULL,
  maxPassages = NULL
)
}
\description{
Adds passage constraints to the given lp, based on the given passage constraints table
}
\examples{
  modules = simulation$modules
  passageCons = simulation$constraints$passage
  minPassages = simulation$control$minPassages
  maxPassages = simulation$control$maxPassages
}
