% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addPrimaryVariables}
\alias{solver.addPrimaryVariables}
\title{Adds primary variables to the given lp, creating a column for each module in the pool}
\usage{
solver.addPrimaryVariables(lp, modules, moduleInf)
}
\description{
Adds primary variables to the given lp, creating a column for each module in the pool
}
\examples{
  modules = simulation$modules
}
