% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addRequiredModulesConstraint}
\alias{solver.addRequiredModulesConstraint}
\title{Adds required module constraints to the given lp, based on the given list of modules already assigned}
\usage{
solver.addRequiredModulesConstraint(lp, modules, reqModuleInd)
}
\description{
Adds required module constraints to the given lp, based on the given list of modules already assigned
}
