% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addTestLengthConstraint}
\alias{solver.addTestLengthConstraint}
\title{Adds test length constraints to the given lp, based on the given test length}
\usage{
solver.addTestLengthConstraint(lp, modules, testLength)
}
\description{
Adds test length constraints to the given lp, based on the given test length
}
