% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.addTIFConstraints}
\alias{solver.addTIFConstraints}
\title{Adds tif constraints to the given lp, based on the given tif table}
\usage{
solver.addTIFConstraints(lp, modules, tif)
}
\description{
Adds tif constraints to the given lp, based on the given tif table
}
\examples{
  modules = simulation$modules
  tif = simulation$tif
}
