% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.disposeProblem}
\alias{solver.disposeProblem}
\title{Destroys the given lp object, cleaning up all associated resources}
\usage{
solver.disposeProblem(lp)
}
\description{
Destroys the given lp object, cleaning up all associated resources
}
