% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.solveAndGetSolution.clp.internal}
\alias{solver.solveAndGetSolution.clp.internal}
\title{Returns the solution for the given lp, solving it internally using the "clpAPI" package Simplex algorithm}
\usage{
solver.solveAndGetSolution.clp.internal(lp.glpk, solverConfig)
}
\description{
Returns the solution for the given lp, solving it internally using the "clpAPI" package Simplex algorithm
}
\examples{
  lp.glpk = initProbGLPK()
  readLPGLPK(lp.glpk, fname="glpkAPI.lp")
  solverConfig = list(name = "clp", external = F, mipGap = 0.0001, timeout = 1000, verbose = T)
  solution.clp = solver.solveAndGetSolution.clp.internal(lp.glpk, solverConfig)
}
