% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.solveAndGetSolution.cplex.internal}
\alias{solver.solveAndGetSolution.cplex.internal}
\title{Returns the solution for the given lp, solving it internally using the "cplexAPI" package
install.packages("cplexAPI", configure.args = "--with-cplex-dir=/path/to/CPLEX_Studio_Community129/cplex", verbose = T)}
\usage{
solver.solveAndGetSolution.cplex.internal(lp.glpk, solverConfig)
}
\description{
Returns the solution for the given lp, solving it internally using the "cplexAPI" package
install.packages("cplexAPI", configure.args = "--with-cplex-dir=/path/to/CPLEX_Studio_Community129/cplex", verbose = T)
}
\examples{
  lp.glpk = initProbGLPK()
  readLPGLPK(lp.glpk, fname="glpkAPI.lp")
  solverConfig = list(name = "cplex", external = F, mipGap = 0.0001, timeout = 1000, verbose = T)
  solution.cplex = solver.solveAndGetSolution.cplex.internal(lp.glpk, solverConfig)
}
