% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.solveAndGetSolution.glpk.external}
\alias{solver.solveAndGetSolution.glpk.external}
\title{Returns the solution for the given lp, solving it externally using the "glpsol" terminal command}
\usage{
solver.solveAndGetSolution.glpk.external(lp.glpk, solverConfig)
}
\description{
Returns the solution for the given lp, solving it externally using the "glpsol" terminal command
}
\examples{
  lp.glpk = initProbGLPK()
  readLPGLPK(lp.glpk, fname="glpkAPI.lp")
  solverConfig = list(name = "glpk", external = T, mipGap = 0.0001, timeout = 1000, verbose = T)
  solution.glpk = solver.solveAndGetSolution.glpk.external(lp.glpk, solverConfig)
}
