% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.solveAndGetSolution.lpsolve.internal}
\alias{solver.solveAndGetSolution.lpsolve.internal}
\title{Returns the solution for the given lp, solving it internally using the "lpSolveAPI" package}
\usage{
solver.solveAndGetSolution.lpsolve.internal(lp.glpk, solverConfig)
}
\description{
Returns the solution for the given lp, solving it internally using the "lpSolveAPI" package
}
\examples{
  lp.glpk = initProbGLPK()
  readLPGLPK(lp.glpk, fname="glpkAPI.lp")
  solverConfig = list(name = "lpsolve", external = F, mipGap = 0.0001, timeout = 1000, verbose = T)
  solution.lpsolve = solver.solveAndGetSolution.lpsolve.internal(lp.glpk, solverConfig)
}
