% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{solver.solveAndGetSolution}
\alias{solver.solveAndGetSolution}
\title{Returns the solution for the given lp, using the solverConfig to choose which tool to solve the problem}
\usage{
solver.solveAndGetSolution(lp.glpk, modules, solverConfig)
}
\description{
Returns the solution for the given lp, using the solverConfig to choose which tool to solve the problem
}
\examples{
  modules = simulation$modules
  solverConfig = simulation$control$solver
}
