% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminationRule.R
\name{tr.csem.isTerminated}
\alias{tr.csem.isTerminated}
\title{Returns whether the given simulee has completed the simulation.}
\usage{
tr.csem.isTerminated(simulee, simulation)
}
\arguments{
\item{simulee}{A data.frame containing at least one simulee, the first row will be used.}

\item{simulation}{An object defining the test that is being run.}
}
\value{
A logical true/false depending on whether the simulee has reached the end of the simulation.
}
\description{
The `csem` algorithm will try to end the test as soon as possible after the
minimum test length has been reached and the csem of the latest item is below
a target value.
}
\examples{
  simulation = readRDS(system.file("example/passage-adaptive-wpm.rds", package = "CATSimulator"))
  simulee = generateSimulees(numSimulees = 1, mean = 0.0, sd = 1.5)
  terminated = tr.csem.isTerminated(simulee, simulation)
}
